requirements.txt
ultralytics>=8.0.0
opencv-python>=4.5.5
numpy>=1.23.0
easyocr>=1.6
streamlit>=1.25.0
pillow>=9.0.0
Flask>=2.0.0
